* some docstring cleanup * moved some tests around
